Change from Raffaele Montuoro:
This commit introduces the ability to set the maximum number of tracers that can be handled by the FMS framework at build time, eliminating the need to edit the FMS source code. This feature is relevant to forecast systems requiring a number of atmospheric tracers larger than FMS hardcoded limit, such as those used for air quality modeling.